Make the suse version check slightly less awful
The version that's being checked here still makes 0 sense

+++ libraries/provider_hab_package.rb
@@ -112,7 +112,7 @@ class Chef
             'target=x86_64-windows'
           elsif platform_family?('rhel') && node['platform_version'].to_i < 6
             'target=x86_64-linux-kernel2'
-          elsif platform_family?('suse') && node['platform_version'].to_f < 6.0
+          elsif platform_family?('suse') && node['platform_version'].to_i < 6
             'target=x86_64-linux-kernel2'
           else
             ''
